feat: Rename 'resume-schema' to 'schema-resume'
Migrated from 'resume-schema' to 'schema-resume' in all the project.
This change includes renaming folders and updating references in index
files, integration tests, and package.json. Reflected this change
consistently across the project for better module structure and
understanding.
