+++ package.json
@@ -4,13 +4,13 @@
 	"type": "module",
 	"exports": {
 		"./resume.yml": "./resume.yml",
-		"./resume-schema": "./src/resume-schema/index.ts",
+		"./schema-resume": "./src/schema-resume/index.ts",
 		"./schema-primitive": "./src/schema-primitive/index.ts"
 	},
 	"publishConfig": {
 		"registry": "https://npm.pkg.github.com",
 		"exports": {
-			"./resume-schema": "./lib/resume-schema/index.js",
+			"./schema-resume": "./lib/schema-resume/index.js",
 			"./schema-primitive": "./lib/schema-primitive/index.js"
 		}
 	},
