+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './resume-schema/index.js'
+export * from './schema-resume/index.js'
 export * from './schema-primitive/index.js'
