+++ src/schema-resume/basics/basic-schema.snapshot.json
