+++ src/schema-resume/certificates/certificate-schema.snapshot.json
