+++ src/schema-resume/education/education-schema.snapshot.json
