+++ src/schema-resume/interest/interest.ts
