+++ src/schema-resume/skill/skill-schema.snapshot.json
