+++ src/schema-resume/skill/skill.ts
